refactor: Delete Service
불필요 Service삭제
