--- app/src/main/java/com/b1nd/alimo/data/remote/service/AlarmService.kt
@@ -1,16 +1,0 @@
-package com.b1nd.alimo.data.remote.service
-
-import android.content.SharedPreferences
-import javax.inject.Inject
-
-class AlarmService @Inject constructor(
-private val sharedPreferences: SharedPreferences
-){
-    fun getAlarmState(): Boolean =
-        sharedPreferences.getBoolean("alarm", false) ?: true
-
-
-    fun setAlarmState(value: Boolean) =
-        sharedPreferences.edit().putBoolean("alarm", value).apply()
-
-}
