+++ app/src/main/java/com/b1nd/alimo/data/repository/FirebaseTokenRepository.kt
@@ -1,22 +1,26 @@
 package com.b1nd.alimo.data.repository
 
+import com.b1nd.alimo.data.Resource
+import com.b1nd.alimo.data.local.dao.FirebaseTokenDao
+import com.b1nd.alimo.data.local.entity.FirebaseTokenEntity
 import com.b1nd.alimo.data.model.FirebaseTokenModel
+import com.b1nd.alimo.data.remote.mapper.toModel
 import com.b1nd.alimo.data.remote.safeFlow
-import com.b1nd.alimo.data.remote.service.FirebaseTokenService
 import javax.inject.Inject
 
 class FirebaseTokenRepository @Inject constructor(
-    private val firebaseTokenService: FirebaseTokenService
+    private val firebaseTokenDao: FirebaseTokenDao
 ){
 
-    suspend fun getToken() = safeFlow<FirebaseTokenModel> {
-        emit(
-            firebaseTokenService.getToken()
-        )
+    suspend fun getToken() =
+        safeFlow<FirebaseTokenModel> {
+        val fcmToken = Resource.Success(firebaseTokenDao.getToken().toModel())
+        emit(fcmToken)
     }
 
-    suspend fun insert(token: String){
-        firebaseTokenService.insert(token)
+    suspend fun insert(fcmToken: String){
+        val fcmToken = FirebaseTokenEntity(fcmToken = fcmToken)
+        firebaseTokenDao.insert(fcmToken)
     }
 
 }
